Refactor GUI components into ui package
