+++ build.gradle
@@ -44,7 +44,7 @@ test {
 }
 
 application {
-    mainClassName = "duke.Duke"
+    mainClassName = "duke.Launcher"
 }
 
 checkstyle {
